Finish Deserializer, remove old serializer, fix numerous std bugs

+++ src/game/scene/world/graph/MeshGenerator.h
@@ -29,7 +29,7 @@ public:
                   std::shared_ptr<BlockChunk> chunk, std::array<std::shared_ptr<BlockChunk>, 6> adjacent,
                   std::array<NoiseSample, 3>& blockOffsets);
 private:
-    BlockDef& getBlockAt(const glm::vec3 &pos);
+    BlockDef& getBlockAt(const glm::ivec3 &pos);
     void addFaces(const glm::vec3 &offset, const std::vector<MeshPart> &meshParts, const glm::vec3& tint);
 
     unsigned int indOffset = 0;
